requirements.txt
Django==5.2.1
matplotlib==3.10.3
numpy==2.2.5
sympy==1.14.0
djangorestframework==3.16.0